refactor: add shared logic to class
